requirements.txt
python==3.8.3

pip==20.0.2

matplotlib==3.1.3

pytorch==1.5.0

torchvision==0.6.0

numpy==1.18.4

opencv-python==4.2.0.32

GitPython==3.1.2

tensorboard==2.2.2

imageio==2.9.0

jupyter==1.0.0

